pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trajectoryrl"
version = "0.1.0"
description = "Bittensor subnet for optimizing AI agent policies using ClawBench"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TrajectoryRL Team"}
]
keywords = ["bittensor", "ai", "agents", "reinforcement-learning", "openclaw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "bittensor>=7.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/trajectoryRL/trajectoryrl"
Documentation = "https://github.com/trajectoryRL/trajectoryrl/blob/main/README.md"
Repository = "https://github.com/trajectoryRL/trajectoryrl"
Issues = "https://github.com/trajectoryRL/trajectoryrl/issues"

[project.scripts]
trajectoryrl-validator = "neurons.validator:main"
trajectoryrl-miner = "neurons.miner:main"

[tool.setuptools.packages.find]
include = ["trajectoryrl*", "neurons*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]  # Line too long (handled by black)
